Cleanup conversion warnings in CMS and some more.

+++ src/main/cms/cms_types.h
@@ -70,11 +70,11 @@ typedef struct
 
 #define IS_PRINTVALUE(p) ((p)->flags & PRINT_VALUE)
 #define SET_PRINTVALUE(p) { (p)->flags |= PRINT_VALUE; }
-#define CLR_PRINTVALUE(p) { (p)->flags &= ~PRINT_VALUE; }
+#define CLR_PRINTVALUE(p) { (p)->flags &= (uint8_t)~PRINT_VALUE; }
 
 #define IS_PRINTLABEL(p) ((p)->flags & PRINT_LABEL)
 #define SET_PRINTLABEL(p) { (p)->flags |= PRINT_LABEL; }
-#define CLR_PRINTLABEL(p) { (p)->flags &= ~PRINT_LABEL; }
+#define CLR_PRINTLABEL(p) { (p)->flags &= (uint8_t)~PRINT_LABEL; }
 
 #define IS_DYNAMIC(p) ((p)->flags & DYNAMIC)
